FIX:좋아요 추가 삭제 구현 완료

+++ src/main/java/com/group3/project_green/heart/HeartController.java
@@ -1,11 +1,13 @@
 package com.group3.project_green.heart;
 
+import com.group3.project_green.Session.SessionUser;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.apache.coyote.Response;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -16,9 +18,13 @@ public class HeartController {
     private final HeartService service;
 
     @PostMapping("")
-    public ResponseEntity<Long> saveHeart(@RequestBody HeartDTO dto) {
+    public ResponseEntity<Long> saveHeart(@RequestBody HeartDTO dto,@AuthenticationPrincipal SessionUser user) {
         log.info("=====================" + dto + "====================");
-        Long hno = service.saveHeart(dto);
+        Heart exist = service.getHeart(user.getId(), dto.getPno());
+        Long hno = null;
+        if (exist == null) {
+            hno = service.saveHeart(dto);
+        }
         return new ResponseEntity<>(hno, HttpStatus.OK);
     }
 
@@ -29,11 +35,23 @@ public class HeartController {
         return new ResponseEntity<>(count, HttpStatus.OK);
     }
 
-    @DeleteMapping("/{hno}/remove")
-    public ResponseEntity<String> remove(@PathVariable("hno") Long hno) {
+    @GetMapping(value = "/{pno}/info", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Heart> getHeartInfo(@PathVariable("pno") Long pno,@AuthenticationPrincipal SessionUser user) {
+        Heart exist = service.getHeart(user.getId(), pno);
+        if(exist != null) {
+            return new ResponseEntity<>(exist, HttpStatus.OK);
+        }else return new ResponseEntity<>(null, HttpStatus.OK);
+    }
+    @DeleteMapping("/{hno}/{pno}/remove")
+    public ResponseEntity<Boolean> remove(@PathVariable("hno") Long hno,@PathVariable("pno") Long pno, @AuthenticationPrincipal SessionUser user) {
         log.info("===================" + hno + "====================");
-        service.remove(hno);
-        return new ResponseEntity<>("success", HttpStatus.OK);
+        boolean result = false;
+        Heart exist = service.getHeart(user.getId(),pno);
+        if(exist != null) {
+            service.remove(hno);
+            result = true;
+        }
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
